updated pipes, decompose components

+++ medicine-delivery-app/server/src/shop/shop.service.ts
@@ -1,17 +1,12 @@
-import { CreateShopDto } from '../dto/createShopDto.dto'
-import { Injectable } from '@nestjs/common'
+import { CreateShopDto } from './dto/createShopDto.dto'
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/mongoose'
 import { Shop } from './shop.model'
-import { Model } from 'mongoose'
-import { CreateOrderDto } from 'src/dto/CreateOrderDto.tdo'
-import { Order } from './order.model'
+import mongoose, { Model } from 'mongoose'
 
 @Injectable()
 export class ShopService {
-	constructor(
-		@InjectModel(Shop.name) private shopModel: Model<Shop>,
-		@InjectModel(Order.name) private orderModel: Model<Order>
-	) {}
+	constructor(@InjectModel(Shop.name) private shopModel: Model<Shop>) {}
 
 	async createShop(dto: CreateShopDto) {
 		const newShop = new this.shopModel(dto)
@@ -19,6 +14,8 @@ export class ShopService {
 	}
 
 	async removeShop(id: string) {
+		const isValidShopId = mongoose.isValidObjectId(id)
+		if (!isValidShopId) throw new HttpException('Invalid shopId', HttpStatus.BAD_REQUEST)
 		return this.shopModel.findByIdAndDelete({ _id: id })
 	}
 
@@ -26,12 +23,12 @@ export class ShopService {
 		return this.shopModel.find({}, 'name _id')
 	}
 	async getProductsByShop(shopId: string) {
+		const isValidObjectId = mongoose.isValidObjectId(shopId)
+		if (!isValidObjectId) throw new HttpException('Invalid shopId', HttpStatus.BAD_REQUEST)
 		const shop = await this.shopModel.findById(shopId).populate('products')
+		if (!shop) {
+			throw new HttpException('Shop was not found', HttpStatus.NOT_FOUND)
+		}
 		return shop.products
-	}
-
-	async createOrder(dto: CreateOrderDto) {
-		const newOrder = new this.orderModel({ ...dto.userData, order: dto.order })
-		return newOrder.save()
 	}
 }
